feat: introduce productsController
