+++ server/src/controllers/index.ts
@@ -1,1 +1,2 @@
 export { default as UsersController } from './usersController'
+export { dafault as ProductsController } from './productsController'
