+++ server/src/index.ts
@@ -1,10 +1,10 @@
-import express from 'express'
-import cors from 'cors'
-import mongoose from 'mongoose'
-require('dotenv').config();
+import express from "express";
+import cors from "cors";
+import mongoose from "mongoose";
+require("dotenv").config();
 
-import { UserRouter } from './routes/user';
-import { ProductRouter } from './routes/product';
+import { UserRouter } from "./routes/user";
+import { ProductRouter } from "./routes/product";
 
 const app = express();
 
@@ -14,6 +14,6 @@ app.use(cors());
 app.use("/user", UserRouter);
 app.use("/products", ProductRouter);
 
-mongoose.connect(`mongodb+srv://pedrolportow:${process.env.DB_PASSWORD}@ecommerce-mern.4yui1uz.mongodb.net/ecommerce-mern`);
+mongoose.connect(`mongodb://0.0.0.0:27017/ecommerce-mern`);
 
-app.listen(3001, () => console.log("Server starterd"))
+app.listen(3001, () => console.log("Server starterd"));
